goreleaser: add homebrew support
This PR adds support for relasing the CLI as a homebrew package for macOS users.

To install the package the command is:

```
brew install planetscale/tap/pscale
```

The homebrew formulas are automatically published to repo https://github.com/planetscale/homebrew-tap. The repo will also allow us to publish other homebrew packages in the future (`brew install planetscale/tap/pscale-proxy`)

We also update the Readme with the new instructions.

closes https://github.com/planetscale/project-big-bang/issues/82