Specify user+password using --defaults-extra-file
If we pass login information to the `mysql` command-line interface using
`-u` and `-p` options, it emits a warning that that's insecure.  And it
is!  The password is visible in the output of `ps`.

This commit changes the `pscale shell` command so it creates a temporary
configuration file with `user=` and `password=` lines in it, then passes
that config file as a `--defaults-extra-file` to the `mysql` CLI.  That
eliminates the warning and the security problem.