cmd: add globally applied -json mode
This removes the local `--json` flags from all subcommands and uses a
**global** `--json` flag. It simplifies the code quite a bit as we don't
have to lookup the local flag anymore and don't have to setup a local
JSON flag for sub commands.

We're going to create followup PR to improve the JSON mode even further
and will display anything printed to stdout/stderr in JSON format
(including user errors). The goal is to make sure `pscale` can be used
in scripts and automation software in an useful way.