cmd: print errors in JSON format if --format is set to JSON
This is a small improvement to change the output of an error, if the `--format` flag is set. Instead of printing it as a single line, we print the error in JSON format if the `--format json` flag is enabled.

Before:
```
$ pscale db list --format json
Error: not authenticated yet. Please run 'pscale auth login'
```

After:

```
$ pscale db list --format json
{"error": "not authenticated yet. Please run 'pscale auth login'"}
```

Btw, I checked other CLI's, and some of them print in human readable format even if you set the format output to JSON, such as `gcloud`.  Hence, I'm OK closing this issue if you think we should output in human readable format.