chmod 0600 access-token
f359a28fda6b changed the default file mode for `access-token` to 0600 (not
readable by anyone except the owner).  That's good.  But any existing file
will still be whatever the old `mode&^umask` was, generally 0664.  That's
bad.

So this change resets the mode on the token file to 0600 if necessary any
time the config is loaded.