Remove vendor/ folder for simplicity
I believe we shouldn't use the vendor folder for the CLI. Once it's open source, people by default (due to the `go` command) will use the official Go module proxy. Every dependency will be cached by default. The Go community has started to embrace this method and `go` works very well nowadays with vendorless projects.

There are many benefits to not having a `vendor/` folder for such a project. One example is not having to deal with a large change set in a Pull Request. Another one is, the user doesn't have to use `go mod vendor`. In overall the `go` command and tools like `gopls` also plays nicer with a vendorless project.

I may be biased, but I once wrote a blog post that goes into more details: https://arslan.io/2019/08/02/why-you-should-use-a-go-module-proxy/