connect: exit if `--execute` finishes
This PR changes the behaviour of the `--execute` flag.

Previously we would continue to serve the proxy even if the command with `--execute` would finish or fail. With this PR we're changing it.  Instead of silently continuing to run, we're exiting completely and shutting down the proxy. I believe this is the correct behaviour given the fact that the user explicitly has passed the `--execute` flag.