cmd: improve execution of the root command
This PR simplifies the execution and printing the error message of the root command. Previously we would print an error message both inside the `cmd/pscale` package and also inside the `internal/cmd` package. On top of this difference, the exit status returned from the root command would also be shadowed from another function call. In our case this was the `updater`.

With this PR, the `cmd/pscale` main package only exits with a given exit code. It's as minimal as its get. By making sure we're also just exiting with a code in the main package, we also set a artifical boundary. This way we don't print any message accidentally inside the `main` package (because all it receives is an integer value).

Next, we seperated the actual execution of the root command and any post function calls. The root command is now called with the `cmd.runCmd()` function. The `cmd.Execute()` is now responsible of alling the root command, printing any useful information/errors to the users, and also making sure to return the correct exit code.

closes: https://github.com/planetscale/cli/issues/283