shell: disable coloring for prompt
This PR disables the custom coloring for the MySQL shell prompt.
Coloring unfortunately causes various issues on different systems (such
as: https://github.com/planetscale/cli/issues/251), and it's  difficult
to debug because we don't have control over the `mysql` command-line
tool.

We used coloring to let the user know that they are using the `main`
branch (caution required). To make sure that the user is connecting to
the `main` branch, we used ASCII characters to distinguish betwen a main
ranch and a non-main branch.

This is how the prompt looks like now for the `main` branch:

```
planetscale/|⚠ main ⚠|> ....
```

closes: https://github.com/planetscale/cli/issues/251