cli: add context cancellation support to all commands
This PR adds proper context cancellation and propagation support to all
commands. Currently, even if the user cancels the operation, we would
continue to make requests to our API.

With the changes introduced in this PR, we now listen to `os.Interrupt`
and `os.Kill` signals. The context associated with the signals are
passed down to all child commands.